+++ src/ts/lib/03-stages/classes/abstract/AbstractStage.ts
@@ -1092,26 +1092,29 @@ export abstract class AbstractStage<
      * 
      * @category Running
      * 
-     * @param subpath  The subdriectory, relative to src path.
-     * @param _distDir  Optionally force a diffrent output directory than the auto-generated one.
+     * @param subpath       The subdriectory, relative to src path.
+     * @param _distDir      Optionally force a diffrent output directory than the auto-generated one.
+     * @param logLevelBase  Base output level for log messages.
      * 
      * @since 0.1.4-alpha
+     * @since 0.2.0-alpha.1 — Added logLevelBase param.
      * 
      * @experimental
      */
     protected async runCustomDirCopySubStage(
         subpath: string,
         _distDir?: string,
+        logLevelBase: number = 1,
     ) {
-        this.console.progress( 'copying ' + subpath + ' to dist...', 1 );
+        this.console.progress( 'copying ' + subpath + ' to dist...', 0 + logLevelBase );
 
         const distDir = _distDir ?? this.getDistDir( undefined ).replace( /\/$/g, '' );
 
         if ( this.fs.exists( distDir ) ) {
-            this.console.verbose( 'deleting any existing files...', 2 );
+            this.console.verbose( 'deleting any existing files...', 1 + logLevelBase );
             this.fs.delete(
                 [ distDir + '/' + subpath ],
-                this.params.verbose ? 3 : 2
+                ( this.params.verbose ? 2 : 1 ) + logLevelBase,
             );
         }
 
@@ -1119,20 +1122,20 @@ export abstract class AbstractStage<
 
         // returns
         if ( !this.fs.exists( srcDir + '/' + subpath ) ) {
-            this.console.verbose( 'ⅹ source dir ' + this.fs.pathRelative( srcDir ) + ' does not exist, exiting...', 2 );
+            this.console.verbose( 'ⅹ source dir ' + this.fs.pathRelative( srcDir ) + ' does not exist, exiting...', 1 + logLevelBase );
             return;
         }
 
         // returns
         if ( !this.fs.isDirectory( srcDir + '/' + subpath ) ) {
-            this.console.verbose( 'ⅹ source dir ' + this.fs.pathRelative( srcDir ) + ' is not a directory, exiting...', 2 );
+            this.console.verbose( 'ⅹ source dir ' + this.fs.pathRelative( srcDir ) + ' is not a directory, exiting...', 1 + logLevelBase );
             return;
         }
 
-        this.console.verbose( 'copying files...', 2 );
+        this.console.verbose( 'copying files...', 1 + logLevelBase );
         this.fs.copy(
             subpath,
-            2,
+            ( this.params.verbose ? 2 : 1 ) + logLevelBase,
             distDir,
             srcDir,
             {
@@ -1152,12 +1155,14 @@ export abstract class AbstractStage<
      * 
      * @category Running
      *
-     * @param subpath   The subdriectory, relative to src path.
-     * @param _distDir  Optionally force a diffrent output directory than the auto-generated one.
-     * @param postCSS   Whether to run PostCSS on the output css. Default true.
+     * @param subpath       The subdriectory, relative to src path.
+     * @param _distDir      Optionally force a diffrent output directory than the auto-generated one.
+     * @param postCSS       Whether to run PostCSS on the output css. Default true.
+     * @param logLevelBase  Base output level for log messages.
      *
      * @since 0.1.4-alpha
      * @since 0.2.0-alpha — Added postCSS param and PostCSS compatibility.
+     * @since 0.2.0-alpha.1 — Added logLevelBase param.
      *
      * @experimental
      */
@@ -1165,27 +1170,28 @@ export abstract class AbstractStage<
         subpath: string,
         _distDir?: string,
         postCSS: boolean = true,
+        logLevelBase: number = 1,
     ) {
-        this.console.progress( 'compiling ' + subpath + ' to css...', 1 );
+        this.console.progress( 'compiling ' + subpath + ' to css...', 0 + logLevelBase );
 
         const distDir = _distDir ?? this.getDistDir( undefined, subpath ).replace( /\/$/g, '' );
 
         if ( this.fs.exists( distDir ) ) {
-            this.console.verbose( 'deleting any existing files...', 2 );
-            this.fs.delete( [ distDir ], this.params.verbose ? 3 : 2 );
+            this.console.verbose( 'deleting any existing files...', 1 + logLevelBase );
+            this.fs.delete( [ distDir ], ( this.params.verbose ? 2 : 1 ) + logLevelBase );
         }
 
         const srcDir = this.getSrcDir( undefined, subpath ).replace( /\/+$/gi, '' );
 
         // returns
         if ( !this.fs.exists( srcDir ) ) {
-            this.console.verbose( 'ⅹ source dir ' + this.fs.pathRelative( srcDir ) + ' does not exist, exiting...', 2 );
+            this.console.verbose( 'ⅹ source dir ' + this.fs.pathRelative( srcDir ) + ' does not exist, exiting...', 1 + logLevelBase );
             return;
         }
 
         // returns
         if ( !this.fs.isDirectory( srcDir ) ) {
-            this.console.verbose( 'ⅹ source dir ' + this.fs.pathRelative( srcDir ) + ' is not a directory, exiting...', 2 );
+            this.console.verbose( 'ⅹ source dir ' + this.fs.pathRelative( srcDir ) + ' is not a directory, exiting...', 1 + logLevelBase );
             return;
         }
 
@@ -1205,12 +1211,12 @@ export abstract class AbstractStage<
 
         // returns
         if ( !scssPaths.length ) {
-            this.console.verbose( 'ⅹ no css, sass, or scss files found', 2 );
+            this.console.verbose( 'ⅹ no css, sass, or scss files found', 1 + logLevelBase );
             return;
         }
 
 
-        this.console.verbose( 'building path arguments...', 2 );
+        this.console.verbose( 'building path arguments...', 1 + logLevelBase );
 
         const scssPathArgs = scssPaths.map(
             ( _path ) => {
@@ -1241,27 +1247,27 @@ export abstract class AbstractStage<
             }
         );
 
-        this.console.vi.debug( { scssPathArgs }, this.params.verbose ? 3 : 2 );
+        this.console.vi.debug( { scssPathArgs }, ( this.params.verbose ? 2 : 1 ) + logLevelBase );
 
 
-        this.console.verbose( 'compiling to css at ' + distDir + '...', 2 );
+        this.console.verbose( 'compiling to css at ' + distDir + '...', 1 + logLevelBase );
         await Promise.all( scssPathArgs.map(
             ( { input, output } ) => this.atry(
                 this.compiler.scss,
-                this.params.verbose ? 3 : 2,
-                [ input, output, this.params.verbose ? 3 : 2 ],
+                ( this.params.verbose ? 2 : 1 ) + logLevelBase,
+                [ input, output, ( this.params.verbose ? 2 : 1 ) + logLevelBase ],
             )
         ) );
 
         if ( postCSS ) {
 
-            this.console.verbose( 'processing with postcss...', 2 );
+            this.console.verbose( 'processing with postcss...', 1 + logLevelBase );
             await this.atry(
                 this.compiler.postCSS,
-                this.params.verbose ? 3 : 2,
+                ( this.params.verbose ? 2 : 1 ) + logLevelBase,
                 [
                     scssPathArgs.map( _o => ( { from: _o.output } ) ),
-                    this.params.verbose ? 3 : 2,
+                    ( this.params.verbose ? 2 : 1 ) + logLevelBase,
 
                 ],
             );